Refactor: rename braille converter to convert braille to text
